Make the exit code 32 bits.
According to Austin Groups bug 594, siginfo_t::si_status should hold the
exit code as passed to exit()/_Exit() entirely -- not just the eight
least significant bits.

+++ src/common/syscalldefs_mi.h
@@ -317,7 +317,7 @@ typedef uint32_t cloudabi_condvar_t;    // pthread_cond_*().
 typedef uint64_t cloudabi_device_t;     // struct stat::st_dev.
 typedef uint64_t cloudabi_dircookie_t;  // readdir().
 typedef uint16_t cloudabi_errno_t;      // errno.
-typedef uint8_t cloudabi_exitcode_t;    // _exit() and _Exit().
+typedef uint32_t cloudabi_exitcode_t;   // _exit() and _Exit().
 typedef uint32_t cloudabi_fd_t;         // File descriptors.
 typedef uint16_t cloudabi_fdflags_t;    // cloudabi_fdstat_t.
 typedef uint16_t cloudabi_fdsflags_t;   // fd_stat_put().
